requirements-dev.txt
-r requirements.txt
black
isort
pytest
flake8
mypy
pytest-cov
pytorch-ignite
pytest-xdist
